Improve test-helpers import path (#1067)
* Improve test-helpers import path

* Still use ember-native-dom-helpers for find and findAll

* Document new import path
