+++ addon-test-support/helpers.js
@@ -1,8 +1,12 @@
-import { deprecate, warn } from '@ember/debug';
+import { deprecate } from '@ember/debug';
 import { registerAsyncHelper } from '@ember/test';
-import wait from 'ember-test-helpers/wait';
-import { click, fillIn, keyEvent, triggerEvent, find, findAll } from 'ember-native-dom-helpers';
-
+import { click, fillIn, keyEvent, triggerEvent, findAll } from 'ember-native-dom-helpers';
+import {
+  selectChoose as _selectChoose,
+  selectSearch as _selectSearch,
+  removeMultipleOption as _removeMultipleOption,
+  clearSelected as _clearSelected
+} from './index';
 /**
  * @private
  * @param {String} selector CSS3 selector of the elements to check the content
@@ -63,169 +67,56 @@ export function touchTrigger() {
 }
 
 export async function selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex) {
-  let trigger, target;
-  if (cssPathOrTrigger instanceof HTMLElement) {
-    if (cssPathOrTrigger.classList.contains('ember-power-select-trigger')) {
-      trigger = cssPathOrTrigger;
-    } else {
-      trigger = find('.ember-power-select-trigger', cssPathOrTrigger);
-    }
-  } else {
-    trigger = find(`${cssPathOrTrigger} .ember-power-select-trigger`);
-
-    if (!trigger) {
-      trigger = find(cssPathOrTrigger);
-    }
-
-    if (!trigger) {
-      throw new Error(`You called "selectChoose('${cssPathOrTrigger}', '${valueOrSelector}')" but no select was found using selector "${cssPathOrTrigger}"`);
-    }
-  }
-
-  if (trigger.scrollIntoView) {
-    trigger.scrollIntoView();
-  }
-
-  let contentId = `${trigger.attributes['aria-owns'].value}`;
-  let content = find(`#${contentId}`);
-  // If the dropdown is closed, open it
-  if (!content || content.classList.contains('ember-basic-dropdown-content-placeholder')) {
-    await click(trigger);
-    await wait();
-  }
-
-  // Select the option with the given text
-  let options = findAll(`#${contentId} .ember-power-select-option`);
-  let potentialTargets = options.filter((opt) => opt.textContent.indexOf(valueOrSelector) > -1);
-  if (potentialTargets.length === 0) {
-    potentialTargets = findAll(`#${contentId} ${valueOrSelector}`);
-  }
-  if (potentialTargets.length > 1) {
-    let filteredTargets = [].slice.apply(potentialTargets).filter((t) => t.textContent.trim() === valueOrSelector);
-    if (optionIndex === undefined) {
-      target = filteredTargets[0] || potentialTargets[0];
-    } else {
-      target = filteredTargets[optionIndex] || potentialTargets[optionIndex];
-    }
-  } else {
-    target = potentialTargets[0];
-  }
-  if (!target) {
-    throw new Error(`You called "selectChoose('${cssPathOrTrigger}', '${valueOrSelector}')" but "${valueOrSelector}" didn't match any option`);
-  }
-  await click(target);
-  return wait();
+  return _selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex);
 }
 
 export async function selectSearch(cssPathOrTrigger, value) {
-  let trigger;
-  if (cssPathOrTrigger instanceof HTMLElement) {
-    trigger = cssPathOrTrigger;
-  } else {
-    let triggerPath = `${cssPathOrTrigger} .ember-power-select-trigger`;
-    trigger = find(triggerPath);
-    if (!trigger) {
-      triggerPath = cssPathOrTrigger;
-      trigger = find(triggerPath);
-    }
-
-    if (!trigger) {
-      throw new Error(`You called "selectSearch('${cssPathOrTrigger}', '${value}')" but no select was found using selector "${cssPathOrTrigger}"`);
-    }
-  }
-
-  if (trigger.scrollIntoView) {
-    trigger.scrollIntoView();
-  }
-
-  let contentId = `${trigger.attributes['aria-owns'].value}`;
-  let isMultipleSelect = !!find('.ember-power-select-trigger-multiple-input', trigger);
-
-  let content = find(`#${contentId}`);
-  let dropdownIsClosed = !content || content.classList.contains('ember-basic-dropdown-content-placeholder');
-  if (dropdownIsClosed) {
-    await click(trigger);
-    await wait();
-  }
-  let isDefaultSingleSelect = !!find('.ember-power-select-search-input');
-
-  if (isMultipleSelect) {
-    await fillIn(find('.ember-power-select-trigger-multiple-input', trigger), value);
-  } else if (isDefaultSingleSelect) {
-    await fillIn('.ember-power-select-search-input', value);
-  } else { // It's probably a customized version
-    let inputIsInTrigger = !!find('.ember-power-select-trigger input[type=search]', trigger);
-    if (inputIsInTrigger) {
-      await fillIn(find('input[type=search]', trigger), value);
-    } else {
-      await fillIn(`#${contentId} .ember-power-select-search-input[type=search]`, 'input');
-    }
-  }
-  return wait();
+  return _selectSearch(cssPathOrTrigger, value);
 }
 
 export async function removeMultipleOption(cssPath, value) {
-  let elem;
-  let items = findAll(`${cssPath} .ember-power-select-multiple-options > li`);
-  let item = items.find((el) => el.textContent.indexOf(value) > -1);
-  if (item) {
-    elem = find('.ember-power-select-multiple-remove-btn', item);
-  }
-  try {
-    await click(elem);
-    return wait();
-  } catch(e) {
-    warn('css path to remove btn not found');
-    throw e;
-  }
+  return _removeMultipleOption(cssPath, value);
 }
 
 export async function clearSelected(cssPath) {
-  let elem = find(`${cssPath} .ember-power-select-clear-btn`);
-  try {
-    await click(elem);
-    return wait();
-  } catch(e) {
-    warn('css path to clear btn not found');
-    throw e;
-  }
+  return _clearSelected(cssPath);
 }
 
 // Helpers for acceptance tests
 export default function() {
   registerAsyncHelper('selectChoose', function(_, cssPathOrTrigger, valueOrSelector, optionIndex) {
     deprecate(
-      'Using the implicit global async helper `selectChoose` is deprecated. Please, import it explicitly with `import { selectChoose } from "ember-power-select/test-support/helpers"`',
+      'Using the implicit global async helper `selectChoose` is deprecated. Please, import it explicitly with `import { selectChoose } from "ember-power-select/test-support"`',
       true,
       { id: 'ember-power-select-global-select-choose', until: '2.0.0' }
     );
-    return selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex);
+    return _selectChoose(cssPathOrTrigger, valueOrSelector, optionIndex);
   });
 
   registerAsyncHelper('selectSearch', async function(app, cssPathOrTrigger, value) {
     deprecate(
-      'Using the implicit global async helper `selectSearch` is deprecated. Please, import it explicitly with `import { selectSearch } from "ember-power-select/test-support/helpers"`',
+      'Using the implicit global async helper `selectSearch` is deprecated. Please, import it explicitly with `import { selectSearch } from "ember-power-select/test-support"`',
       true,
       { id: 'ember-power-select-global-select-search', until: '2.0.0' }
     );
-    return selectSearch(cssPathOrTrigger, value);
+    return _selectSearch(cssPathOrTrigger, value);
   });
 
   registerAsyncHelper('removeMultipleOption', async function(app, cssPath, value) {
     deprecate(
-      'Using the implicit global async helper `removeMultipleOption` is deprecated. Please, import it explicitly with `import { removeMultipleOption } from "ember-power-select/test-support/helpers"`',
+      'Using the implicit global async helper `removeMultipleOption` is deprecated. Please, import it explicitly with `import { removeMultipleOption } from "ember-power-select/test-support"`',
       true,
       { id: 'ember-power-select-global-remove-multiple-option', until: '2.0.0' }
     );
-    return removeMultipleOption(cssPath, value);
+    return _removeMultipleOption(cssPath, value);
   });
 
   registerAsyncHelper('clearSelected', async function(app, cssPath) {
     deprecate(
-      'Using the implicit global async helper `clearSelected` is deprecated. Please, import it explicitly with `import { clearSelected } from "ember-power-select/test-support/helpers"`',
+      'Using the implicit global async helper `clearSelected` is deprecated. Please, import it explicitly with `import { clearSelected } from "ember-power-select/test-support"`',
       true,
       { id: 'ember-power-select-global-clear-selected', until: '2.0.0' }
     );
-    return clearSelected(cssPath);
+    return _clearSelected(cssPath);
   });
 }
