+++ package.json
@@ -55,7 +55,7 @@
     "ember-cli-htmlbars-inline-precompile": "^1.0.2",
     "ember-cli-inject-live-reload": "^1.7.0",
     "ember-cli-mirage": "^0.4.0",
-    "ember-cli-qunit": "^4.2.0",
+    "ember-cli-qunit": "^4.3.2",
     "ember-cli-sass": "^7.1.0",
     "ember-cli-template-lint": "0.7.5",
     "ember-cli-uglify": "^2.0.0",
