+++ tests/dummy/app/templates/snippets/test-helpers-10-js.js
@@ -1,0 +1,1 @@
+import { selectChoose, selectSearch, removeMultipleOption, clearSelected } from 'ember-power-select/test-support';
